Re-org contracts
