+++ contracts/stellar/Cargo.toml
@@ -1,6 +1,28 @@
-[package]
-name = "stellar"
-version = "0.1.0"
-edition = "2021"
+[workspace]
+resolver = "2"
 
-[dependencies]
+members = [
+    "attestation",
+    "authority",
+    "schema",
+]
+
+[workspace.package]
+version = "0.0.0"
+
+[workspace.dependencies]
+soroban-sdk = { version = "22.0.7" }
+
+[profile.release]
+opt-level = "z"
+overflow-checks = true
+debug = 0
+strip = "symbols"
+debug-assertions = false
+panic = "abort"
+codegen-units = 1
+lto = true
+
+[profile.release-with-logs]
+inherits = "release"
+debug-assertions = true
