+++ contracts/stellar/attestation/src/lib.rs
@@ -1,4 +1,5 @@
-use soroban_sdk::{contract, contractimpl, Address, Env, Vec};
+#![no_std]
+use soroban_sdk::{contract, contractimpl, Address, Env, String, Vec};
 
 #[derive(Debug, Clone)]
 pub struct Attestation {
@@ -54,15 +55,13 @@ impl AttestationContract {
         let caller = self.principal();
 
         // Check if caller is a registered authority
-        let authority = get_authority(env, caller)
-            .ok_or("Caller is not a registered authority")?;
+        let authority = get_authority(env, caller).ok_or("Caller is not a registered authority")?;
 
         // Check if schema UID is provided
         let schema_uid = ref_uid.ok_or("Schema UID not provided")?;
 
         // Retrieve the schema from storage
-        let schema = get_schema(env, &schema_uid)
-            .ok_or("Schema not found")?;
+        let schema = get_schema(env, &schema_uid).ok_or("Schema not found")?;
 
         // Process levy if applicable
         if let Some(levy) = &schema.levy {
@@ -76,7 +75,7 @@ impl AttestationContract {
         let uid_bytes = uid.to_vec();
 
         // Ensure attestation does not already exist
-        if fetch_attestation(env, &uid_bytes).is_some() {
+        if get_attestation(env, &uid_bytes).is_some() {
             return Err("Attestation with this UID already exists".to_string());
         }
 
@@ -97,6 +96,34 @@ impl AttestationContract {
 
         Ok(())
     }
+
+    fn revoke_attestation(&self, env: Env, uid: Vec<u8>) -> Result<(), String> {
+        let caller = self.principal();
+
+        // Fetch the attestation from storage
+        let mut attestation = get_attestation(env, &uid).ok_or("Attestation not found")?;
+
+        // Ensure the caller is the original attester
+        if attestation.attester != caller {
+            return Err("Only the original attester can revoke the attestation".to_string());
+        }
+
+        // Update the attestation's revocation_time
+        attestation.revocation_time = Some(env.block_timestamp());
+
+        // Store the updated attestation
+        store_attestation(env, &uid, &attestation)?;
+
+        // Emit a revocation event
+        emit_attestation_revoked(env, uid);
+
+        Ok(())
+    }
+
+    fn query_attestation(&self, env: Env, uid: Vec<u8>) -> Result<Option<Attestation>, String> {
+        let attestation = get_attestation(env, &uid).ok_or("Attestation not found")?;
+        Ok(Some(attestation))
+    }
 }
 
 // Storage access functions
@@ -108,7 +135,7 @@ fn get_schema(env: Env, uid: &[u8]) -> Option<Schema> {
     env.storage().get(uid)
 }
 
-fn fetch_attestation(env: Env, uid: &[u8]) -> Option<Attestation> {
+fn get_attestation(env: Env, uid: &[u8]) -> Option<Attestation> {
     env.storage().get(uid)
 }
 
@@ -116,4 +143,11 @@ fn store_attestation(env: Env, uid: &[u8], attestation: &Attestation) -> Result<
     env.storage()
         .set(uid, attestation)
         .map_err(|_| "Failed to store attestation".to_string())
+}
+
+// Event Emission
+fn emit_attestation_revoked(env: Env, uid: Vec<u8>) {
+    env.events()
+        .publish("AttestationRevoked", (uid,))
+        .expect("Failed to emit attestation revoked event");
 }
