+++ contracts/stellar/authority/Makefile
@@ -1,0 +1,16 @@
+default: build
+
+all: test
+
+test:
+	cargo test
+
+build:
+	stellar contract build
+	@ls -l target/wasm32-unknown-unknown/release/*.wasm
+
+fmt:
+	cargo fmt --all
+
+clean:
+	cargo clean
